Almost ready - fix unit tests

+++ scripts/tests/compose_and_test_selected_apps.sh
@@ -2,7 +2,39 @@
 
 OSV_DIR=$(readlink -f $(dirname $0)/../..)
 CAPSTAN_REPO=$HOME/.capstan
+
+argv0=${0##*/}
+usage() {
+	cat <<-EOF
+	Compose and test apps out ouf pre-built capstan packages
+
+	Usage: ${argv0} [options] <group_of_tests> | <test_app_name>
+	Options:
+	  -c              Compose test app image only
+	  -r              Run test app only (must have been composed earlier)
+	  -R              Compose test app image with RoFS (ZFS is the default)
+	EOF
+	exit ${1:-0}
+}
+
 FS=zfs
+COMPOSE_ONLY=false
+RUN_ONLY=false
+
+while getopts crRh: OPT ; do
+	case ${OPT} in
+	c) COMPOSE_ONLY=true;;
+	r) RUN_ONLY=true;;
+	R) FS=rofs;;
+	h) usage;;
+	?) usage 1;;
+	esac
+done
+
+shift $((OPTIND - 1))
+[[ -z $1 ]] && usage 1
+
+TEST="$1"
 
 compose_test_app()
 {
@@ -10,22 +42,34 @@ compose_test_app()
   local DEPENDANT_PKG1=$2
   local DEPENDANT_PKG2=$3
 
-  local DEPENDENCIES="--require osv.$APP_NAME"
-  if [ "$DEPENDANT_PKG1" != "" ]; then
-     DEPENDENCIES="--require osv.$DEPENDANT_PKG1 $DEPENDENCIES"
-  fi
-  if [ "$DEPENDANT_PKG2" != "" ]; then
-     DEPENDENCIES="--require osv.$DEPENDANT_PKG2 $DEPENDENCIES"
+  local IMAGE_PATH="$CAPSTAN_REPO/repository/test-$APP_NAME/test-$APP_NAME.qemu"
+
+  if [ $RUN_ONLY == false ]; then
+    local DEPENDENCIES="--require osv.$APP_NAME"
+    if [ "$DEPENDANT_PKG1" != "" ]; then
+      DEPENDENCIES="--require osv.$DEPENDANT_PKG1 $DEPENDENCIES"
+    fi
+    if [ "$DEPENDANT_PKG2" != "" ]; then
+      DEPENDENCIES="--require osv.$DEPENDANT_PKG2 $DEPENDENCIES"
+    fi
+
+    echo "-----------------------------------"
+    echo " Composing $APP_NAME ... "
+    echo "-----------------------------------"
+
+    TEMPDIR=$(mktemp -d) && pushd $TEMPDIR > /dev/null && \
+      capstan package compose $DEPENDENCIES --fs $FS "test-$APP_NAME" && \
+      rmdir $TEMPDIR && popd > /dev/null
+  else
+    echo "Reusing the test image: $IMAGE_PATH that must have been composed before!"
   fi
 
-  echo "-----------------------------------"
-  echo " Composing $APP_NAME ... "
-  echo "-----------------------------------"
-
-  TEMPDIR=$(mktemp -d) && pushd $TEMPDIR > /dev/null && \
-    capstan package compose $DEPENDENCIES --fs $FS "test-$APP_NAME" && \
-    rmdir $TEMPDIR && popd > /dev/null
-  cp $CAPSTAN_REPO/repository/"test-$APP_NAME"/"test-$APP_NAME".qemu $OSV_DIR/build/release/usr.img
+  if [ -f "$IMAGE_PATH" ]; then
+    cp $CAPSTAN_REPO/repository/"test-$APP_NAME"/"test-$APP_NAME".qemu $OSV_DIR/build/release/usr.img
+  else
+    echo "Could not find test image: $IMAGE_PATH!" 
+    exit 1
+  fi
 }
 
 run_test_app()
@@ -33,14 +77,16 @@ run_test_app()
   local OSV_APP_NAME=$1
   local TEST_PARAMETER=$2
 
-  echo "-----------------------------------"
-  echo " Testing $OSV_APP_NAME ... "
-  echo "-----------------------------------"
+  if [ $COMPOSE_ONLY == false ]; then
+    echo "-----------------------------------"
+    echo " Testing $OSV_APP_NAME ... "
+    echo "-----------------------------------"
 
-  if [ -f $OSV_DIR/apps/$OSV_APP_NAME/test.sh ]; then
-    $OSV_DIR/apps/$OSV_APP_NAME/test.sh $TEST_PARAMETER
-  elif [ -f $OSV_DIR/modules/$OSV_APP_NAME/test.sh ]; then
-    $OSV_DIR/modules/$OSV_APP_NAME/test.sh $TEST_PARAMETER
+    if [ -f $OSV_DIR/apps/$OSV_APP_NAME/test.sh ]; then
+      $OSV_DIR/apps/$OSV_APP_NAME/test.sh $TEST_PARAMETER
+    elif [ -f $OSV_DIR/modules/$OSV_APP_NAME/test.sh ]; then
+      $OSV_DIR/modules/$OSV_APP_NAME/test.sh $TEST_PARAMETER
+    fi
   fi
   echo ''
 }
@@ -103,12 +149,8 @@ run_unit_tests()
   compose_test_app "httpserver-api-https-tests" "httpserver-api-tests" && run_test_app "httpserver-api" "https"
 }
 
-if [ "$1" == "" ]; then
-  echo "Usage: $0 <test_type> OR <test_app_name>"
-  exit 1
-fi
 
-case "$1" in
+case "$TEST" in
   simple)
     echo "Testing simple apps ..."
     echo "-----------------------------------"
@@ -128,12 +170,12 @@ case "$1" in
   all)
     echo "Running all tests ..."
     echo "-----------------------------------"
+    run_unit_tests
     test_simple_apps
     test_http_apps
-    test_apps_with_tester
-    run_unit_tests;;
+    test_apps_with_tester;;
   *)
-    compose_and_run_test_app $1
+    compose_and_run_test_app $TEST
 esac
 
 #
